refactor: consolidate test directories and update documentation
- Merge test/service/ and test/mock-integration/ → test/contract/
- Rename integration-harness.js → handler-harness.js
- Replace verbose AGENTS.md with concise contract test guide
- Focus on testEventHandler usage patterns and practical examples
- All tests continue to pass with reorganized structure
