+++ test/contract/cogni-evaluated-gates-behavior.test.js
