+++ test/contract/legacy-spec-bug.test.js
